\name{alligator.dat}
\alias{alligator.dat}
\title{Data frame of alligator food choice by lake, gender and size.}
\description{
  A 5x4x2x2 table of alligators food choice, lake, gender and size
  stored in a data frame.
}
\usage{data(alligator.dat)
}
\format{\code{allgiator.dat} is a data frame
}
\references{Agresti (1990). \emph{Categorical Data
  Analysis}. Wiley-Interscience. Page 308.
}
\keyword{datasets}
