\name{czech.dat}
\alias{czech.dat}
\docType{data}
\title{Czech auto workers data}
\description{
  Czech auto workers data
}
\usage{data(czech.dat)}
\format{
  A data frame with 64 observations on the following 7 variables.
  \describe{
    \item{y}{a numeric vector}
    \item{A}{a factor with levels \code{no} \code{yes}}
    \item{B}{a factor with levels \code{no} \code{yes}}
    \item{C}{a factor with levels \code{no} \code{yes}}
    \item{D}{a factor with levels \code{small} \code{large}}
    \item{E}{a factor with levels \code{small} \code{large}}
    \item{F}{a factor with levels \code{neg} \code{pos}}
  }
}
\source{
  Edwards and Havranek (1985)
}
\examples{
data(czech.dat)
}
\keyword{datasets}
