\name{titanic.dat}
\alias{titanic.dat}
\docType{data}
\title{Titanic Survival Data}
\description{
  A data frame of counts of titanic passengers classified
  by class, age, sex and survival
}
\usage{data(titanic.dat)}
\format{
  A data frame with 32 observations on the following 5 variables.
  \describe{
    \item{y}{a numeric vector}
    \item{class}{a numeric vector}
    \item{age}{a numeric vector}
    \item{sex}{a numeric vector}
    \item{surv}{a numeric vector}
  }
}
\source{
  Cytel web site http://www.cytel.com
}
\examples{
data(titanic.dat)
}
\keyword{datasets}
