\name{hyper}
\alias{hyper}
\title{Generalized hypergeometric distribution on the log scale}
\description{
  Returns the log hypergeometric conditional probability of a vector of Poisson
  counts up to an additive constant of proportionallity.
}
\usage{
hyper(y)
}
\arguments{
  \item{y}{Counts}
}
\details{
}
\value{
  A real-valued scalar.
}
\references{}
\author{Brian S. Caffo}
\note{}
\seealso{\code{\link{mcexact}}}

\examples{
}
\keyword{htest}
